Cut and paste most formatting logic to a new class
